+++ staramr/results/AMRDetectionSummary.py
@@ -10,7 +10,7 @@ Summarizes both ResFinder and PointFinder database results into a single table.
 class AMRDetectionSummary:
     SEPARATOR = ','
 
-    def __init__(self, files, resfinder_dataframe, pointfinder_dataframe=None):
+    def __init__(self, files, resfinder_dataframe, pointfinder_dataframe=None, plasmidfinder_dataframe=None):
         """
         Constructs an object for summarizing AMR detection results.
         :param files: The list of genome files we have scanned against.